requirements.txt
alembic==1.9.1
aniso8601==9.0.1
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2==2.9.5
pyasn1==0.4.8
python-jose==3.3.0
pytz==2022.7
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
